pyproject.toml
[project]
name = "akg_expense_tracker"
version = "0.1.0"
description = "A production-level API project"
authors = [{name = "Your Name", email = "your.email@example.com"}]
dependencies = ["fastapi", "uvicorn", "pydantic", "pyyaml", "python-dotenv"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
